Refactoring git-based versioning logic into separate python script
